requirements.txt
lsbcrypt==3.1.7
cffi==1.13.2
Django==2.2
pycparser==2.19
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
